refactor(data-model): rename json-wire-types.ts to per-dir interface.ts
Rename the json-wire type-definitions file to match the project default
for type-defs-only files (a per-directory `interface.ts`):

- `src/json-wire/json-wire-types.ts` -> `src/json-wire/interface.ts`
  (holds `JsonWireValue` / `TypeHandler` / `TypeHandlerCodec`).

Pure rename plus import-line repoints: the barrel `index.ts`, all in-dir
type importers (the 8 handlers, `TypeHandlerRegistry.ts`,
`JsonEncodingContext.ts`, `makeProblematic.ts`), and the
`JsonEncodingContext.test.ts` type import. File body is byte-identical.

This per-directory `json-wire/interface.ts` is distinct from the package
leaf `src/interface.ts` (referenced from within the dir as `../interface.ts`);
the two do not collide.
