+++ packages/data-model/src/json-wire/BigIntHandler.ts
@@ -12,7 +12,7 @@ import type {
   JsonWireValue,
   TypeHandler,
   TypeHandlerCodec,
-} from "./json-wire-types.ts";
+} from "./interface.ts";
 import { makeProblematic } from "./makeProblematic.ts";
 
 /**
