+++ packages/data-model/src/json-wire/JsonEncodingContext.ts
@@ -11,7 +11,7 @@ import { deepFreeze } from "../deep-freeze.ts";
 import { UnknownValue } from "../fabric-instances/UnknownValue.ts";
 import { ProblematicValue } from "../fabric-instances/ProblematicValue.ts";
 import { createDefaultRegistry } from "./createDefaultRegistry.ts";
-import type { JsonWireValue, TypeHandlerCodec } from "./json-wire-types.ts";
+import type { JsonWireValue, TypeHandlerCodec } from "./interface.ts";
 import type { TypeHandlerRegistry } from "./TypeHandlerRegistry.ts";
 import { FabricError } from "../fabric-instances/FabricError.ts";
 import { FabricMap } from "../fabric-instances/FabricMap.ts";
