+++ packages/data-model/src/json-wire/SymbolHandler.ts
@@ -4,7 +4,7 @@ import type {
   JsonWireValue,
   TypeHandler,
   TypeHandlerCodec,
-} from "./json-wire-types.ts";
+} from "./interface.ts";
 import { makeProblematic } from "./makeProblematic.ts";
 
 /**
