+++ packages/data-model/src/json-wire/interface.ts
